abstract deleted, merged into single command file
